+++ packages/eslint-plugin-pf-codemods/test/rules/tab-title-text.js
@@ -4,7 +4,7 @@ const rule = require('../../lib/rules/tab-title-text');
 ruleTester.run("tab-title-text", rule, {
   valid: [
     {
-      code: `import { Tab, TabTitleText } from '@patternfly/react-core'; <Tab title={<TabTitleText>Title</TabTitleText>}>Content</Tab>`,
+      code: `import { Tab, TabTitleText, TabTitleIcon } from '@patternfly/react-core'; <Tab title={<TabTitleText>Title</TabTitleText>}>Content</Tab>`,
     },
     {
       // No @patternfly/react-core import
@@ -15,14 +15,26 @@ ruleTester.run("tab-title-text", rule, {
     {
       code:   `import { Tab } from '@patternfly/react-core';
 <Tab title="Title">Content</Tab>;
-<Tab title="Title">Content</Tab>;`,
-      output: `import { Tab, TabTitleText } from '@patternfly/react-core';
+<Tab title="Title">Content</Tab>;
+<Tab title={<UsersIcon />}>Content</Tab>;
+<Tab title={<><UsersIcon /> Text</>}>Content</Tab>;`,
+      output: `import { Tab, TabTitleText, TabTitleIcon } from '@patternfly/react-core';
 <Tab title={<TabTitleText>Title</TabTitleText>}>Content</Tab>;
-<Tab title={<TabTitleText>Title</TabTitleText>}>Content</Tab>;`,
+<Tab title={<TabTitleText>Title</TabTitleText>}>Content</Tab>;
+<Tab title={<TabTitleIcon><UsersIcon /></TabTitleIcon>}>Content</Tab>;
+<Tab title={<><TabTitleIcon><UsersIcon /></TabTitleIcon><TabTitleText> Text</TabTitleText></>}>Content</Tab>;`,
       errors: [
         {
-          message: 'add missing imports TabTitleText from @patternfly/react-core',
+          message: 'add missing imports TabTitleText, TabTitleIcon from @patternfly/react-core',
           type: "ImportDeclaration"
+        },
+        {
+          message: `title needs to be wrapped with the TabTitleText and/or TabTitleIcon component`,
+          type: "JSXOpeningElement",
+        },
+        {
+          message: `title needs to be wrapped with the TabTitleText and/or TabTitleIcon component`,
+          type: "JSXOpeningElement",
         },
         {
           message: `title needs to be wrapped with the TabTitleText and/or TabTitleIcon component`,