+++ packages/pf-codemods/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.15](https://github.com/patternfly/pf-codemods/compare/pf-codemods@1.0.14...pf-codemods@1.0.15) (2020-05-29)
+
+**Note:** Version bump only for package pf-codemods
+
+
+
+
+
+## [1.0.14](https://github.com/patternfly/pf-codemods/compare/pf-codemods@1.0.13...pf-codemods@1.0.14) (2020-05-29)
+
+**Note:** Version bump only for package pf-codemods
+
+
+
+
+
+## [1.0.13](https://github.com/patternfly/pf-codemods/compare/pf-codemods@1.0.12...pf-codemods@1.0.13) (2020-05-28)
+
+**Note:** Version bump only for package pf-codemods
+
+
+
+
+
+## [1.0.12](https://github.com/patternfly/pf-codemods/compare/pf-codemods@1.0.11...pf-codemods@1.0.12) (2020-05-28)
+
+**Note:** Version bump only for package pf-codemods
+
+
+
+
+
 ## [1.0.11](https://github.com/patternfly/pf-codemods/compare/pf-codemods@1.0.10...pf-codemods@1.0.11) (2020-05-28)
 
 **Note:** Version bump only for package pf-codemods